third_party/WebKit/LayoutTests/http/tests/inspector/resource-tree/resource-tree-no-xhrs-expected.txt

CONSOLE WARNING: line 98: Setting 'XMLHttpRequest.withCredentials' for synchronous requests is deprecated.
CONSOLE MESSAGE: line 6: XHR loaded: 1
Tests that XHRs are not added to resourceTreeModel.

Bug 60321
